refactor: split advisor system probing
